+++ src/components/modules/Image.vue
@@ -50,12 +50,16 @@ export default {
   methods: {
     init () {
       this.$database.view(this.data.type, this.data.mediaid, (res) => {
-        if (res.resource.service === 'upload') {
-          this.$firebase.imagesRef.child(res.resource._id).getDownloadURL().then((url) => {
-            res.resource.thumbURL = url
-            res.resource.imageURL = url
-            res.resource.pageURL = url
-          })
+        if (!this.$root.user.nqUser) {
+          if (res.resource.service === 'upload') {
+            this.$firebase.imagesRef.child(res.resource._id).getDownloadURL().then((url) => {
+              res.resource.thumbURL = url
+              res.resource.imageURL = url
+              res.resource.pageURL = url
+            })
+          }
+        } else {
+          console.log('pull nq resource...')
         }
         this.media = res.resource
       })